consolidate base template modernizr code
Currently multiple cf.gov base page templates include Modernizr.js (and
use it to disable JS) in different ways. This PR consolidates this logic
and unifies it across the templates. This change is motivated by, and
also removes, reference to a deprecated common.ie9.js file that was
removed in 9afbc47.

The three base templates involved in this change are:

1) cfgov/jinja2/v1/_layouts/base-common.html: This is a Jinja2 template
rendered by Django with the Jinja2 template engine. This is used as the
base template for most, if not all, of the Wagtail page templates.

2) cfgov/legacy/templates/front/base_nonresponsive.html: This is a
Django template rendered by Django with the Django template engine. This
is used by some pages in cfgov-refresh (for example, the agreements app
at https://www.consumerfinance.gov/credit-cards/agreements/) as well as
pages from satellite apps (for example, the Paying for College page at
https://www.consumerfinance.gov/paying-for-college/compare-financial-aid-and-college-cost/).

3) cfgov/legacy/templates/front/base_update.html: This is a Django
template rendered by Django with the Django template engine. This is not
used by any pages in cfgov-refresh but is used by satellite apps (for
example, the retirement app at
https://www.consumerfinance.gov/consumer-tools/retirement/before-you-claim/).

This change unifies the way that the Modernizr code is loaded into each
of these templates. Before, the Jinja2 template loaded the code directly
with an {% include %} tag, and the Django templates loaded with a
<script> tag. Now, both embed the code directly.

The code exists only once now, as a Jinja template snippet at
cfgov/v1/jinja/v1/modernizr.html, alongside other similar snippets for
the site header and footer. All of those snippets are now exposed to
the Django template engine through a base_elements templatetag library
(formerly named header_footer). Old references and unit tests have all
been updated.

Some other minor changes in this PR: I've removed the custom Modernizr
build for Spanish Ask CFPB answer pages instead of using the global one.
This change also makes it so that all of the pages include the Modernizr
script at the end of the <head> section, instead of in the <body> tag.

To test, run a local server with the retirement satellite app installed.
Visit these pages to test each template:

http://localhost:8000/
http://localhost:8000/credit-cards/agreements/
http://localhost:8000/consumer-tools/retirement/before-you-claim/

When accessing these pages in IE9, you should see a "no-js" class on the
page <html> tag.